pyproject.toml
[project]
name = "wireframe-distillation"
version = "0.1.0"
description = ""
authors = [{ name = "Iván Ferre"}]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
  "torch>=1.12.0",
  "torchvision>=0.13.0",
  "kornia>=0.6.5",
  "numpy>=1.23.0",
  "pillow>=8.3.0",
  "opencv-python>=4.5.0",
  "matplotlib",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
